refactor(daemon): Rename terminator to context, the file
